changes update routine to improve performance

+++ static_aid/DataExtractor_ArchivesSpace.py
@@ -14,7 +14,17 @@ class DataExtractor_ArchivesSpace(DataExtractor):
         lastExport = self.lastExportTime()
         self.makeDestinations()
         headers = self.authenticate()
-        self.findObjectsById(headers)
+        updated_list = self.getUpdatedRefIds(lastExport, headers)
+        for directory in os.listdir(config.assets['src']):
+            id_list = self.getIdList(directory)
+            if len(id_list) > 0:
+                self.getNewObjects(directory, id_list, headers)
+            if len(set(updated_list) & set(id_list)) > 0:
+                logging.info("Exporting updated objects")
+                for ref_id in set(updated_list) & set(id_list):
+                    self.findObjectById(directory, ref_id, headers)
+                    id_list.remove(ref_id)
+
         # self.findResources(lastExport, headers)
         # self.findObjects(lastExport, headers)
         # self.findAgents(lastExport, headers)
@@ -53,7 +63,6 @@ class DataExtractor_ArchivesSpace(DataExtractor):
             logging.info('*** Getting a list of resources modified since %d ***', lastExport)
         else:
             logging.info('*** Getting a list of all resources ***')
-
         url = '%s/resources?all_ids=true&modified_since=%d' % (config.archivesSpace['repository_url'], lastExport)
         resourceIds = requests.get(url, headers=headers)
         for resourceId in resourceIds.json():
@@ -100,23 +109,68 @@ class DataExtractor_ArchivesSpace(DataExtractor):
             else:
                 self.removeFile(objectId, config.destinations['objects'])
 
-    # Looks for objects based on a list of refids
-    def findObjectsById(self, headers):
-        for directory in os.listdir(config.assets['src']):
-            for dirpath, dirnames, filenames in os.walk('%s/%s' % (config.assets['src'], directory)):
-                id_list = []
-                for filename in [f for f in filenames if not f.startswith('.')]:
-                    id_list.append(os.path.basename(os.path.splitext(filename)[0]))
-                for objectId in id_list:
-                    url = '%s/find_by_id/archival_objects?ref_id[]=%s' % (config.archivesSpace['repository_url'], str(objectId))
-                    results = requests.get(url, headers=headers).json()
-                    for result in results["archival_objects"]:
-                        result_url = '%s%s' % (config.archivesSpace['base_url'], result['ref'])
-                        archival_object = requests.get(result_url, headers=headers).json()
-                        if archival_object["publish"]:
-                            self.saveFile(objectId, archival_object, config.destinations[directory])
-                        else:
-                            self.removeFile(objectId, config.destinations[directory])
+    # gets a list of filenames without extensions
+    def getIdList(self, directory):
+        logging.info('*** Getting a list of all filenames without extensions in '+directory+'***')
+        id_list = []
+        for dirpath, dirnames, filenames in os.walk('%s/%s' % (config.assets['src'], directory)):
+            for filename in [f for f in filenames if not f.startswith('.')]:
+                id_list.append(os.path.basename(os.path.splitext(filename)[0]))
+        return id_list
+
+    # gets JSON for updated description
+    def getUpdatedObjects(self, directory, lastExport, id_list, headers):
+        if lastExport > 0:
+            logging.info('*** Getting a list of objects modified since %d ***', lastExport)
+        else:
+            logging.info('*** Getting a list of all objects ***')
+        url = '%s/archival_objects?all_ids=true&modified_since=%d' % (config.archivesSpace['repository_url'], lastExport)
+        archival_objects = requests.get(url, headers=headers)
+        for objectId in archival_objects.json():
+            url = '%s/archival_objects/%s' % (config.archivesSpace['repository_url'], str(objectId))
+            archival_object = requests.get(url, headers=headers).json()
+            if (archival_object["ref_id"] in id_list):
+                if archival_object["publish"]:
+                    self.saveFile(objectId, archival_object, config.destinations[directory])
+                else:
+                    self.removeFile(objectId, config.destinations[directory])
+                id_list.remove(archival_object["ref_id"])
+        return id_list
+
+    # gets JSON for updated description
+    def getUpdatedRefIds(self, lastExport, headers):
+        updated_list = []
+        if lastExport > 0:
+            logging.info('*** Getting a list of ref ids for objects modified since %d ***', lastExport)
+        else:
+            logging.info('*** Getting a list of ref ids for all objects ***')
+        url = '%s/archival_objects?all_ids=true&modified_since=%d' % (config.archivesSpace['repository_url'], lastExport)
+        archival_objects = requests.get(url, headers=headers)
+        for objectId in archival_objects.json():
+            url = '%s/archival_objects/%s' % (config.archivesSpace['repository_url'], str(objectId))
+            archival_object = requests.get(url, headers=headers).json()
+            updated_list.append(archival_object["ref_id"])
+        return updated_list
+
+    # gets JSON for newly added assets
+    def getNewObjects(self, directory, id_list, headers):
+        logging.info("**** Getting data for newly added assets in "+directory+" ***")
+        for objectId in id_list:
+            if not(os.path.isfile(os.path.join(config.DATA_DIR, directory, objectId+'.json'))):
+                self.findObjectById(directory, objectId, headers)
+
+    # Gets JSON for an object by ref_id
+    def findObjectById(self, directory, objectId, headers):
+        url = '%s/find_by_id/archival_objects?ref_id[]=%s' % (config.archivesSpace['repository_url'], str(objectId))
+        results = requests.get(url, headers=headers).json()
+        for result in results["archival_objects"]:
+            result_url = '%s%s' % (config.archivesSpace['base_url'], result['ref'])
+            archival_object = requests.get(result_url, headers=headers).json()
+            if archival_object["publish"]:
+                logging.info("Exporting "+objectId)
+                self.saveFile(objectId, archival_object, config.destinations[directory])
+            else:
+                self.removeFile(objectId, config.destinations[directory])
 
     # Looks for agents
     def findAgents(self, lastExport, headers):
